requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
transformers>=4.35.0
torch>=2.1.0
lxml>=4.9.0
brotli>=1.1.0
